CommunityExpertCL/requirements.txt
torch>=2.0
torch_geometric>=2.4
numpy
PyYAML
ogb
scikit-learn
scipy
tqdm
matplotlib
